research-agent/backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
crewai>=0.80.0
tavily-python>=0.3.0
anthropic>=0.28.0
python-dotenv>=1.0.0
